Allow registration of LearningTasks.
This CL adds the ability to register a LearningTask with a
LearningSessionImpl.  It, in turn, creates a LearningTaskController
for that task.

The LearningTaskController is responsible for taking training
examples, and forwarding them to learner(s), along with feature
selection and accuracy reporting.

Change-Id: I071dd08063b1ff2482ca4e9170f7f2d0cb6bd10a
Reviewed-on: https://chromium-review.googlesource.com/c/1315934
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#609127}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d1d183b221a3c1bcf62b4296666a360ca511daf7