media/filters: Relax the vp9 colorspace matching restriction
For the Intra only frames in profile0, we assign the default
color_space :BT_601 as per the spec. But there are lazily encoded
streams with UNKNOWN color space information in non-intra-only frames
and that will cause enum comparison mismatch for the current frame header
with the frame in the reference list. So we relax the requirement
a bit to cover more streams similar to the other media stacks like
libvpx, gstreamer etc

BUG=889768
R=dalecurtis@chromium.org, hiroh@chromium.org

Change-Id: I5c343d03a941a3aa1447d4f0d3cd4979b92c4adb
Reviewed-on: https://chromium-review.googlesource.com/c/1342776
Reviewed-by: Fredrik Hubinette <hubbe@chromium.org>
Commit-Queue: Fredrik Hubinette <hubbe@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#609929}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cecd07c128d36d28b12046878c44cda3c8b6e09e