Turn on new RemotePlayback by default
This CL turns on the new RemotePlayback pipeline by default.

There are still some issues to address, but none of them are blocking.
Due to the relatively small use of this feature, it is best to turn the
feature on by default, to get the coverage we need. The extended period
of time between M72 branch cut and M72 stable cut will be used to merge
back any unforeseen regressions that pop-up.

NOTE: This change might lead to regressions or improvements of various
tests.

Bug: 790766
Change-Id: I3e2ed2073da6a28b5ec6cb1ee6d2dd1851df1fde
Reviewed-on: https://chromium-review.googlesource.com/c/1338887
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#613277}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f6bc9ac810f69e3310afdb522343447aeee9ea45