[media] Fix race between ready and dismiss in VdaVideoDecoder.
During the time between when VdaVideoDecoder posts PictureReady to the
parent thread and when that task runs, it is possible for a VDA to
dismiss the associated picture buffer.

This CL adds a hop through the parent thread for DismissPictureBuffer().

Bug: 906623
Change-Id: I46db8aff8b7687e5ec7b00c70068002fe2f2983e
Reviewed-on: https://chromium-review.googlesource.com/c/1362005
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Ted Meyer <tmathmeyer@chromium.org>
Commit-Queue: Dan Sanders <sandersd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#615265}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 35b82bc8409d987903dae8ff366ebebb9d526bf0