media/gpu/veaunittest: Initialize Ozone in global environment
The Ozone initialization and creation of the rendering thread was
performed before running each individual tests. This triggers a
CalledOnValidSequence() DCHECK in Debug builds when
AfterSandboxEntry() tries to start the DRM thread for the second time,
and Thread::Start() is called from a different thread that with the
first test.

Fix this by moving all Ozone initialization and the rendering thread to
the global test environment.

BUG=907767
TEST=Checked that VEA was passing on Cheza with --single-process-tests.

Change-Id: Ic9b52a31310b40c99368620ea78cf861d084032d
Reviewed-on: https://chromium-review.googlesource.com/c/1369755
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#616633}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 78b2ef02daff55860c6c6ff70c880a6095f02e46