media/gpu: Add MEDIA_GPU_EXPORT to ImageProcessor
MEDIA_GPU_EXPORT is required so that ImageProcessor is used outside of
libmedia_gpu.so. A test for ImageProcessor is one of such examples. This is also
required if we separate ImageProcessor part from
VideoDecoder/VideoDecodeAccelerator and perform it in their client.

TBR=posciak@chromium.org, acourbot@chromium.org

Bug: 917951
Test: Compile chorme
Change-Id: I5ea16bf3fd6d1dcde04cf71226f290304f50aa41
Reviewed-on: https://chromium-review.googlesource.com/c/1390784
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#619025}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aa719808f3e9c8db99598b2c451c857d3d84f206