media: Use NullMediaLog in tests and most places instead of MediaLog
Currently MediaLog (implementation) can be constructed directly and
be used wherever a MediaLog (interface) is needed. This is widely used
in tests which is fine, and in some production code, which is
questionable (see Bug for details). Part of the problem is that MediaLog
is both an interface and an implementation. And it's not obvious to
people not familiar with the code that MediaLog (implementation)
actually doesn't report the log anywhere.

This CL uses NullMediaLog as much as possible to make it clear it
doesn't actually report the logs.

In the next CL, we'll make sure MediaLog cannot be constructed directly
to prevent this kind of misuse. In the future, we probably should
separate MediaLog interface and implementation.

TBR=dcheng@chromium.org,halliwell@chromium.org

Bug: 712310
Test: No functionality change.
Change-Id: I57c21fb5e4a0a2cba34d6f7d4fabd2d5261fe725
Reviewed-on: https://chromium-review.googlesource.com/c/1391859
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#619191}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 14b3166d786b77a72c2fd5f1c844ba5c8b3b6209