APM in AS: Explicitly set ExperimentalAgc to false when not enabled
It turns out enabling the experimental AGC is the default, so not
explicitly disabling it if not wanted will leave it on.

This mirrors the changes in [1] but for the APM integration used when
running the APM in the audio process.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1393340

Bug: 918677
Change-Id: Iaad99218b6d5cbaa7c9bf7a7342521222440ab40
Reviewed-on: https://chromium-review.googlesource.com/c/1402793
Reviewed-by: Max Morin <maxmorin@chromium.org>
Commit-Queue: Oskar Sundbom <ossu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#621538}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c066ec3dae113e6379b9009acab172cb5283eba4