Remove duplicate close in FakeJpegDecodeAccelerator
Running content_browsertests --gtest_filter=VideoCaptureBrowserTest.ReceiveFramesFromFakeCaptureDevice/19
produced log outputs complaining about trying to close an invalid file handle.
It appears the cause is that a SharedMemoryHandle is closed manually right after
having handed off ownership of it to a WritableUnalignedMapping.

This CL removes the manual call to close. It seems to have been added
by mistake in
https://chromium-review.googlesource.com/c/chromium/src/+/1117689

Note that the test case is currently disabled on Linux due to flakiness.
The issue fixed in this CL is probably not the cause for the flakiness.

Bug: 921694
Change-Id: I158fd243d4a62af6d786ab36a12ca74e04056de7
Reviewed-on: https://chromium-review.googlesource.com/c/1419260
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Christian Fremerey <chfremer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#623926}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 003e572af5c0935ebc68007945cae07b895a61db