MSE-BufferByPts(andByDts): Enforce non-negative timestamp assumptions in SBS::Append
Once FrameProcessor has a frame that has survived processing (e.g.,
timestampOffset application, satisfying any needs-random-access-point
condition for the frame's track, append-window trimming and filtering,
and failing parse for resulting negative DTS if buffering-by-DTS), it
enqueues such a frame for later SourceBufferStream::Append. If it is the
first frame that FrameProcessor thinks may begin a new coded frame
group, it also signals SourceBufferStream::OnStartOfNewCodedFrameGroup
with the frame's DTS and PTS.

The net result is that two conditions (basically that the DTS or PTS,
depending on runtime buffering model, is non-negative in the result of
OnStartOfNewCodedFrameGroup processing and for the first frame in the
queue of frames in an Append), for which SourceBufferStream would
MEDIA_LOG and emit a parse error if they were not met, should never
fail.

This change makes those conditions CHECKS instead of parse errors, to
enable fuzzing to help verify those two conditions are invariant.

BUG=922893

Change-Id: Icae9fdc908943103b8ffea717c35060deeaef0d4
Reviewed-on: https://chromium-review.googlesource.com/c/1423357
Commit-Queue: Matthew Wolenetz <wolenetz@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#624352}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9963e505940bcc41f290b19c29a18b4030e80fbe