media/gpu/v4l2: move picture clearing back to the correct thread
CL crrev.com/c/1391649 changed the thread on which picture clearing
takes place from the child thread to the decode thread. This can result
in a picture being sent before an EGLImage is created for it, since the
later event happens on the child thread and is scheduled before the
picture clearing. Moving picture clearing to the decoder thread removes
any order guarantee that we had before.

Revert this part to what it was before crrev.com/c/1391649 since it
seems to be a typo.

Bug: 917310
Bug: 920908
Test: VDA passing in debug mode on nyan_big
Change-Id: Ic44c37f6aaa6a966c2f11a5e7363975229ec04a8
Reviewed-on: https://chromium-review.googlesource.com/c/1424629
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#624526}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a2062b754caeca0f1ce708b686ee91a0fff2bad4