Fix initialization in //media pipeline integration fuzzers on Mac.
The issue is that because of the aggressive dead code stripping
LLVMFuzzerInitialize may be stripped out by linker on Mac.

We tried to fix that in a universal way (see https://crbug.com/754124
and https://crbug.com/787723), but ended up with a decision that use of
LLVMFuzzerInitialize is so rare that we should just add some visibility
attribute to the function declaration, when it is really needed. See
https://crrev.com/c/834457 for an example.

Bug: 920924
Change-Id: I821cb8bed757fae960e937bfd8501d0c1e5b0038
Reviewed-on: https://chromium-review.googlesource.com/c/1435438
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: Max Moroz <mmoroz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#625861}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 58f4e54141e7c19604f3b69eab13c222382e57f8