Remove DC component from BeepingSource
The current generator is maintaining a value of 0 when silent and
varying between 0 and 128 when beeping. As it generates unsigned 8 bit
audio, this means it's stuck at the lowest value when silent, rather
than at the mid-point. This has confused a test.

Bug: 892534
Change-Id: I4bf5d676ef58e0443716213a5650baacf3f11b8c
Reviewed-on: https://chromium-review.googlesource.com/c/1430074
Reviewed-by: Max Morin <maxmorin@chromium.org>
Commit-Queue: Oskar Sundbom <ossu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#626103}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7a7b211a99eb05f5f3cd5f6f16ddc8fa542bc8d2